// packages/datadog-api-client-v2/models/SecurityMonitoringRuleQuery.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SecurityMonitoringRuleQueryAggregation } from './SecurityMonitoringRuleQueryAggregation';
import { SecurityMonitoringRuntimeAgentRule } from './SecurityMonitoringRuntimeAgentRule';
/**
* Query for matching rule.
*/
export declare class SecurityMonitoringRuleQuery {
    'agentRule'?: SecurityMonitoringRuntimeAgentRule;
    'aggregation'?: SecurityMonitoringRuleQueryAggregation;
    /**
    * Field for which the cardinality is measured. Sent as an array.
    */
    'distinctFields'?: Array<string>;
    /**
    * Fields to group by.
    */
    'groupByFields'?: Array<string>;
    /**
    * The target field to aggregate over when using the sum or max aggregations.
    */
    'metric'?: string;
    /**
    * Name of the query.
    */
    'name'?: string;
    /**
    * Query to run on logs.
    */
    'query'?: string;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): SecurityMonitoringRuleQuery;
    static serialize(data: SecurityMonitoringRuleQuery): {
        [key: string]: any;
    };
    constructor();
}